refactor: 拆分 TodoDrawer / SessionManager / KanbanBoard 为多文件
- TodoDrawer.tsx: 提取 ExecutorPicker, PromptEditor, SkillSelector,
  SchedulerSection, TemplateModal, constants 到 todo-drawer/
- SessionManager.tsx: 提取 helpers, StatsCards, SessionDetailDrawer 到 sessions/
- KanbanBoard.tsx: 提取 constants, helpers 到 kanban/
